refactor(MoveTimesValidator, InputView): separate move times input validation
