+++ src/main/java/util/MoveTimesValidator.java
@@ -1,0 +1,18 @@
+package util;
+
+public class MoveTimesValidator {
+	private MoveTimesValidator() {
+	}
+
+	public static void validate(String moveTimes) {
+		validateMoveTimes(moveTimes);
+	}
+
+	private static void validateMoveTimes(String moveTimes) {
+		try {
+			Integer.parseInt(moveTimes);
+		} catch (NumberFormatException e) {
+			throw new IllegalArgumentException();
+		}
+	}
+}
